+++ src/pages/Home.jsx
@@ -1,0 +1,17 @@
+import "../common.css"
+import "./styles/Home.css"
+import HomeBoard from "../components/Home/Board"
+
+function HomePage() {
+  return (
+    <>
+      <div className="col">
+        <h1> Chess </h1>
+        <br />
+        <HomeBoard/>
+      </div>
+    </>
+  )
+}
+
+export default HomePage